Create Particle Variables for zone related values.
Allow the checking of zone related target and current temperature.

Particle Variables don't have a `float` type, so switch these to
`double` to make them usable.

+++ src/zone.h
@@ -17,10 +17,10 @@ class Zone
     String m_name;
     int m_pin;
 
-    float m_current_temp;
+    double m_current_temp;
     system_tick_t m_last_current_temp;
 
-    float m_target_temp;
+    double m_target_temp;
 
     bool m_heating;
     system_tick_t m_last_state_change;
